+++ R/zzz.R
@@ -1,5 +1,7 @@
 
 .onLoad <- function(...){
-    requireNamespace("ggplot2")
+  requireNamespace("ggplot2")
+  requireNamespace("compiler")
+  compiler::enableJIT(3)
   return(NULL)
 }